Allow specifying a pattern for the tests to be run
When running on Android, it's quite a pain to select as subset of the
tests using the selector.  Hence, allow a "pattern" URL query
parameter to select which tests to be run.  The key is "pattern" and
the value is a regular expression which is used to match the name of
the test.  If the name matches, the test is selected.